+++ src/main/java/duke/Duke.java
@@ -2,6 +2,10 @@ package duke;
 
 import duke.command.Command;
 import duke.exception.DukeException;
+import duke.util.Parser;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 
 /**
  * The Duke program is a chatbot named Beep Boop Bot that
