+++ src/main/java/duke/command/SortCategoryCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 
 /**
  * The SortCategoryCommand is the command for sorting the tasks in the task list by their categories.
