+++ src/main/java/duke/gui/MainWindow.java
@@ -1,5 +1,8 @@
-package duke;
+package duke.gui;
 
+import duke.Duke;
+import duke.gui.DialogBox;
+import duke.util.Ui;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
