+++ src/test/java/duke/ParserTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
+import duke.util.Parser;
 import org.junit.jupiter.api.Test;
 
 import duke.command.AddDeadlineCommand;
