+++ src/test/java/duke/TaskListTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.time.LocalDate;
 import java.util.ArrayList;
 
+import duke.util.TaskList;
 import org.junit.jupiter.api.Test;
 
 import duke.task.Deadline;
