--- text-ui-test/input.txt
@@ -1,9 +1,0 @@
-todo borrow book
-deadline return book /by Sunday
-event project meeting /from Mon 2pm /to 4pm
-list
-mark 2
-list
-unmark 2
-list
-bye
